Refactor graph crate to have a nicer API Closes #68
