+++ src/graphs/algorithms/dijkstra.rs
@@ -1,6 +1,6 @@
 use super::super::super::algorithms::priority_queue::PriorityQueue;
-use crate::graphs::datastructures::digraph::{Digraph, NodeIndex};
-use crate::graphs::datastructures::graph_path::GraphPath;
+use crate::graphs::datastructures::GraphPath;
+use crate::graphs::datastructures::{Digraph, NodeIndex};
 
 #[cfg(test)]
 #[path = "dijkstra_tests.rs"]
