+++ src/graphs/output/mod.rs
@@ -1,2 +1,5 @@
-pub mod graphviz;
-pub mod svg;
+mod graphviz;
+mod svg;
+
+pub use self::graphviz::*;
+pub use self::svg::*;
